requirements.txt
absl-py==1.4.0
alabaster==0.7.13
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
decorator==4.4.2
docutils==0.20.1
esbonio==0.16.1
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.12.2
flatbuffers==23.5.26
future==0.18.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
imagesize==1.4.1
ipdb==0.13.13
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.6
llvmlite==0.40.1
lsprotocol==2023.0.0a2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
openai-whisper==20231106
opt-einsum==3.3.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.23.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pygls==1.0.2
Pygments==2.15.1
pyspellchecker==0.7.2
redis==4.6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
snowballstemmer==2.2.0
sounddevice==0.4.6
SpeechRecognition==3.10.0
Sphinx==7.1.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spotipy==2.23.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
traitlets==5.9.0
typeguard==3.0.2
typing_extensions==4.5.0
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
